+++ tests/ReplicateFilesystemAdapterTest.php
@@ -13,8 +13,11 @@ use League\Flysystem\FileAttributes;
 use League\Flysystem\FilesystemAdapter;
 use League\Flysystem\InMemory\InMemoryFilesystemAdapter;
 use League\Flysystem\UrlGeneration\PublicUrlGenerator;
+use League\Flysystem\UrlGeneration\TemporaryUrlGenerator;
 use League\Flysystem\Visibility;
 use LogicException;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 
 final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
 {
@@ -54,7 +57,7 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
     /**
      * @return array<array{0: string, 1: array<mixed>, 2: bool, 3?: mixed}>
      */
-    public function callProvider()
+    public static function callProvider()
     {
         return [
             ['fileExists', ['path'], false, true],
@@ -77,12 +80,11 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
     }
 
     /**
-     * @dataProvider callProvider
-     * @test
-     *
      * @param array<mixed> $arguments
      * @param mixed $return
      */
+    #[DataProvider('callProvider')]
+    #[Test]
     public function method_delegation(string $method, array $arguments, bool $useReplica, $return = null): void
     {
         $source = $this->createMock(FilesystemAdapter::class);
@@ -107,25 +109,19 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
         $this->assertSame($return, \call_user_func_array([$this->adapter(), $method], $arguments));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function getting_source_adapter(): void
     {
         $this->assertSame(static::$source, $this->adapter()->getSourceAdapter());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function getting_replica_adapter(): void
     {
         $this->assertSame(static::$replica, $this->adapter()->getReplicaAdapter());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function deleting_file_from_replica(): void
     {
         $source = new InMemoryFilesystemAdapter();
@@ -143,9 +139,7 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
         $this->adapter()->delete('path');
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function writing_non_seekable_stream(): void
     {
         stream_wrapper_register('test', NonSeekableStream::class);
@@ -166,60 +160,50 @@ final class ReplicateFilesystemAdapterTest extends FilesystemAdapterTestCase
         stream_wrapper_unregister('test');
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function generating_a_public_url(): void
     {
-        if (!is_callable('parent::generating_a_public_url')) {
+        if (!interface_exists(PublicUrlGenerator::class)) {
             $this->markTestSkipped();
         }
         static::$adapter = new ReplicateFilesystemAdapter(new PublicUrlGeneratorAdapter(), static::$replica);
         parent::generating_a_public_url();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function generating_a_temporary_url(): void
     {
-        if (!is_callable('parent::generating_a_temporary_url')) {
+        if (!interface_exists(TemporaryUrlGenerator::class)) {
             $this->markTestSkipped();
         }
         static::$adapter = new ReplicateFilesystemAdapter(new TemporaryUrlGeneratorAdapter(), static::$replica);
         parent::generating_a_temporary_url();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function get_checksum(): void
     {
-        if (!is_callable('parent::get_checksum')) {
+        if (!interface_exists(ChecksumProvider::class)) {
             $this->markTestSkipped();
         }
         static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);
         parent::get_checksum();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function cannot_get_checksum_for_non_existent_file(): void
     {
-        if (!is_callable('parent::cannot_get_checksum_for_non_existent_file')) {
+        if (!interface_exists(ChecksumProvider::class)) {
             $this->markTestSkipped();
         }
         static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);
         parent::cannot_get_checksum_for_non_existent_file();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function cannot_get_checksum_for_directory(): void
     {
-        if (!is_callable('parent::cannot_get_checksum_for_directory')) {
+        if (!interface_exists(ChecksumProvider::class)) {
             $this->markTestSkipped();
         }
         static::$adapter = new ReplicateFilesystemAdapter(new ChecksumProviderAdapter(), static::$replica);